Changed file hierarchy
and moved validators to diffrent files
