+++ flask-server/database_stuff/app.py
@@ -18,10 +18,9 @@ def create_app():
     
     db.init_app(app)
 
-    from views import views
-    from auth import auth
-
-    from addProp import addProp
+    from views.views_container import views
+    from views.auth_container import auth
+    from views.addProp import addProp
     
     app.register_blueprint(views, url_prefix='/')
     app.register_blueprint(auth, url_prefix='/')
