+++ flask-server/database_stuff/service/companyService.py
